us-workflow-server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-dotenv==1.0.1
loguru==0.7.3
pydantic==2.10.4
